# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Proxy server for GREEN-API WhatsApp methods.",
        "title": "GREEN-API Integration",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/green/instance/{idInstance}/send-file/{apiToken}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Send a file by URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "idInstance",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API token",
                        "name": "apiToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SendFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/green/instance/{idInstance}/send-message/{apiToken}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Send a text message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "idInstance",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API token",
                        "name": "apiToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/green/instance/{idInstance}/settings/{apiToken}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Get instance settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "idInstance",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API token",
                        "name": "apiToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/green/instance/{idInstance}/state/{apiToken}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Get instance state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "idInstance",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API token",
                        "name": "apiToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "raw": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handler.SendFileRequest": {
            "type": "object",
            "required": [
                "chatId",
                "urlFile"
            ],
            "properties": {
                "caption": {
                    "type": "string"
                },
                "chatId": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "urlFile": {
                    "type": "string"
                }
            }
        },
        "handler.SendMessageRequest": {
            "type": "object",
            "required": [
                "chatId",
                "message"
            ],
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}